replace all of test functions in alone file
