+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = shellenv
-version = 1.3
+version = 1.4
 author = Andrew Rzhavskov
 author_email = rjaan@yandex.ru
 description = print SHELL's enviroment variables on standard output
