+++ src/shellenv/__init__.py
@@ -2,14 +2,16 @@
 from typing import Any, Optional
 
 from .shellenv import ( unsetenv, setenv, getenv, nmaxhitsenv, nminhitsenv, changenv, print_all )
+from .testrun import ( functionality_check )
 
-all = [ 'unsetenv', 
-        'setenv', 
-        'getenv', 
-        'nmaxhitsenv', 
-        'nminhitsenv', 
+all = [ 'unsetenv',
+        'setenv',
+        'getenv',
+        'nmaxhitsenv',
+        'nminhitsenv',
         'changenv',
-        'print_all', ]
+        'print_all', 
+        'functionality_check' ]
 
 # eof
